% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_eig_df}
\alias{generate_eig_df}
\title{Generate a sample dataset from a set of random eigenvalues}
\usage{
generate_eig_df(k = 501, beta = 0.7, nobs = 25, fn)
}
\arguments{
\item{k}{Total number of columns generated, i.e. number of columns of X + 1. Default=501}

\item{beta}{Expected proportion of eigenvalues that are 0. Default=0.7}

\item{nobs}{Number of observations in generated dataset}

\item{fn}{Function to generate a single eigenvalue}
}
\description{
Generate a sample dataset from a set of random eigenvalues
}
