% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_condition_number}
\alias{get_condition_number}
\title{Calculate the condition number, either based on the raw eigenvalues, or based on the regularized ones.
Note: Currently the regularized method only supports the 'minmax' regularization method.}
\usage{
get_condition_number(hmc_res, regularized = F, reg_method = "minmax")
}
\arguments{
\item{hmc_res}{A hmc result objects containing eig_values and tau entries}

\item{regularized}{Should the condition value be calculated based on regularized eigenvalues}

\item{reg_method}{If regularized eigenvalues are used, which method has been used for regularization}
}
\description{
Calculate the condition number, either based on the raw eigenvalues, or based on the regularized ones.
Note: Currently the regularized method only supports the 'minmax' regularization method.
}
