% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_esjd}
\alias{get_esjd}
\title{Calculate Squared Jumping Distance}
\usage{
get_esjd(hmc_res)
}
\arguments{
\item{hmc_res}{A hmc result objects containing energy and samples entries.}
}
\description{
Calculate Squared Jumping Distance
}
