% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_reg_eig}
\alias{get_reg_eig}
\title{Get regularized eigenvalues
Note: Currently the regularized method only supports the 'minmax' regularization method.}
\usage{
get_reg_eig(hmc_res, reg_method = "minmax")
}
\arguments{
\item{hmc_res}{A hmc result objects containing eig_values and tau entries.}

\item{reg_method}{If regularized eigenvalues are used, which method has been used for regularization}
}
\description{
Get regularized eigenvalues
Note: Currently the regularized method only supports the 'minmax' regularization method.
}
