% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{leapfrog_integration}
\alias{leapfrog_integration}
\title{Leapfrog integrator}
\usage{
leapfrog_integration(
  current_state,
  momentum,
  step_size,
  gradient_function,
  num_steps
)
}
\arguments{
\item{current_state}{A number.}

\item{momentum}{A number.}

\item{step_size}{A number.}

\item{gradient_function}{A number.}

\item{num_steps}{A number.}
}
\value{
A list of proposed state and proposed momentum
}
\description{
Leapfrog integrator
}
