% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_eigenvalues}
\alias{plot_eigenvalues}
\title{Plot eigenvalue spectra for each step of the adaptation}
\usage{
plot_eigenvalues(hmc_res, log = F, margin = F)
}
\arguments{
\item{hmc_res}{A list containing the entry 'eig_values' normally output from the hamiltonian_mcmc function when metric_method is set to something that returns eigenvalues.}

\item{log}{Should the plot be drawn on log scale?}

\item{margin}{Should a marginal density of the lower eigenvalue treshold be included?}
}
\description{
Plot eigenvalue spectra for each step of the adaptation
}
