% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_eigenvalues_minmax_shrink}
\alias{plot_eigenvalues_minmax_shrink}
\title{Plot eigenvalue spectra for each step of the adaptation.
Use this alternative when minimum/maximum truncation/shrinking of eigenvalues has been applied.}
\usage{
plot_eigenvalues_minmax_shrink(hmc_res, log = F)
}
\arguments{
\item{hmc_res}{A list containing the entry 'eig_values' normally output from the hamiltonian_mcmc function when metric_method is set to something that returns eigenvalues.}

\item{log}{Should the plot be drawn on log scale?}
}
\description{
Plot eigenvalue spectra for each step of the adaptation.
Use this alternative when minimum/maximum truncation/shrinking of eigenvalues has been applied.
}
