% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_esjd}
\alias{plot_esjd}
\title{Plot Expected Squared Jumping Distance}
\usage{
plot_esjd(hmc_lst)
}
\arguments{
\item{hmc_res_list}{A list of hmc result objects}
}
\description{
Plot Expected Squared Jumping Distance
}
