% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_esjd_potential}
\alias{plot_esjd_potential}
\title{Plot ESJD over Proposed Potential Energy}
\usage{
plot_esjd_potential(hmc_res)
}
\arguments{
\item{hmc_res}{A hmc result object.}
}
\description{
Plot ESJD over Proposed Potential Energy
}
