% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_esjd_trace}
\alias{plot_esjd_trace}
\title{Plot ESJD over Sample Covariance Matrix Trace}
\usage{
plot_esjd_trace(hmc_lst)
}
\arguments{
\item{hmc_lst}{A names list of hmc result objects.}
}
\description{
Plot ESJD over Sample Covariance Matrix Trace
}
