% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ess}
\alias{plot_ess}
\title{Plot Effective Sample Size / second}
\usage{
plot_ess(hmc_res_list)
}
\arguments{
\item{hmc_res_list}{A list of hmc result objects}
}
\description{
Plot Effective Sample Size / second
}
