% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ess_var}
\alias{plot_ess_var}
\title{Plot Effective Sample Size per variable}
\usage{
plot_ess_var(hmc_res_list)
}
\arguments{
\item{hmc_res_list}{A list of hmc result objects}
}
\description{
Plot Effective Sample Size per variable
}
