% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_leapfrog_steps}
\alias{plot_leapfrog_steps}
\title{Plot Leapfrog integration steps}
\usage{
plot_leapfrog_steps(hmc_res, s = NULL, d = NULL)
}
\arguments{
\item{hmc_res}{A list containing the entries 'samples', 'leapfrog_states', normally output from the hamiltonian_mcmc function when returnleapfrogdetails=TRUE.}

\item{s}{Which specific sample to plot. If NULL, a random sample will be plotted.}

\item{d}{Which two variables will be plotted, given as a vector of two integers. If NULL, the two dimensions will be chosen at random.}
}
\description{
Plot Leapfrog integration steps
}
