% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_posterior_draws}
\alias{plot_posterior_draws}
\title{Get regularized eigenvalues
Note: Currently the regularized method only supports the 'minmax' regularization method.}
\usage{
plot_posterior_draws(
  hmc_lst,
  nsamples = 250,
  i = sample(ncol(hmc_lst$Spectral$samples), 1)
)
}
\arguments{
\item{hmc_lst}{A named list of hmc result objects.}

\item{nsamples}{The function will plot the posterior for the last 'nsamples' samples.}

\item{i}{Specifies the variable index for which to generate the plot. Default: Randomly sampled from the available variables.}
}
\description{
Get regularized eigenvalues
Note: Currently the regularized method only supports the 'minmax' regularization method.
}
