% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_step_size}
\alias{plot_step_size}
\title{Plot Effective Step Size}
\usage{
plot_step_size(hmc_res_list)
}
\arguments{
\item{hmc_res_list}{A list of hmc result objects}
}
\description{
Plot Effective Step Size
}
