% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_tau}
\alias{plot_tau}
\title{Plot the adaptation of parameters tau and tau_2}
\usage{
plot_tau(hmc_res)
}
\arguments{
\item{hmc_res}{A list containing the entry 'tau' normally output from the hamiltonian_mcmc function when metric_method is set to something that returns eigenvalues.}
}
\description{
Plot the adaptation of parameters tau and tau_2
}
