% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot Trace (sum of diagonal/eigenvalues of sample covariance matrix)}
\usage{
plot_trace(hmc_lst)
}
\arguments{
\item{hmc_lst}{A list of hmc result objects.}
}
\description{
Plot Trace (sum of diagonal/eigenvalues of sample covariance matrix)
}
