% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regularize_eigvals}
\alias{regularize_eigvals}
\title{Get regularized eigenvalues
Note: Currently the regularized method only supports the 'minmax' regularization method.}
\usage{
regularize_eigvals(eig_vals, tau_min, tau_max, reg_method = "minmax")
}
\arguments{
\item{eig_vals}{Eigenvalues}

\item{tau_min}{tau_min}

\item{tau_max}{tau_max}

\item{reg_method}{If regularized eigenvalues are used, which method has been used for regularization}
}
\description{
Get regularized eigenvalues
Note: Currently the regularized method only supports the 'minmax' regularization method.
}
