% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{leapfrog_integration}
\alias{leapfrog_integration}
\title{Hamiltonian Markov Chain Monte Carlo sampler.}
\usage{
leapfrog_integration(
  current_state,
  momentum,
  step_size,
  stan_obj,
  num_steps,
  returndetails = F,
  M
)
}
\arguments{
\item{current_state}{A number.}

\item{momentum}{A number.}

\item{step_size}{A number.}

\item{stan_obj}{A number.}

\item{num_steps}{A number.}
}
\value{
A list of proposed state, proposed momentum and final gradient
}
\description{
`hamiltonian_mcmc()` is the main function of this library, that provides a simple way to run a Hamiltonian MCMC sampler for given density and gradient functions.

At the moment, no vignette exists.
Leapfrog integrator
}
