% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_esjd_trace}
\alias{plot_esjd_trace}
\title{Plot ESJD over Proposed Potential Energy}
\usage{
plot_esjd_trace(hmc_lst)
}
\arguments{
\item{hmc_res}{A hmc result object.}
}
\description{
Plot ESJD over Proposed Potential Energy
}
